fix waitpid and organize files
