+++ Makefile
@@ -16,7 +16,7 @@ LIBFT_A = $(LIBFT_DIR)/libft.a
 SRC_DIR = main
 EXECUTOR = executor handle_pwd handle_cd handle_echo handle_env handle_export handle_unset pipex
 INIT = init
-PARSER = assign_types parser process_args
+PARSER = assign_types parser process_args skip_redirects expander
 PRINTS = print
 UTILS = utils utils2 utils3 split split_utils cmds error_handle
 SIGNALS = signals
