+++ src/executor/pipex.c
@@ -1,9 +1,7 @@
 #include "minishell.h"
 
-static void	parent_process(int pid, int *new_fd)
+static void	parent_process(int *new_fd)
 {
-	(void)pid;
-	(void)new_fd;
  	dup2(new_fd[0], STDIN_FILENO);
 	close(new_fd[0]);
 	close(new_fd[1]);
@@ -11,28 +9,35 @@ static void	parent_process(int pid, int *new_fd)
 
 static void	prepare_exec(t_tokens *tokens, t_shell *shell)
 {
-	int	res;
-	char **cmds;
+	int		res;
+	char	**cmds;
+	char	*path;
 
 	cmds = put_cmds(tokens);
 	res = ft_isbuiltin(tokens, shell);
 	if (res)
 	{
-		free_array(cmds, arr_len(cmds));
+		free_paths(cmds);
 		exit(0);
 	}
-	if (execve(get_path(tokens->token, shell->envp), cmds, NULL) == -1)
-		perror("error execve");
+	path = get_path(tokens->token, shell->envp);
+	if (!path || execve(path, cmds, NULL) == -1)
+	{
+		free_paths(cmds);
+		if (path)
+			free(path);
+		exit(10);
+	}
+	free_paths(cmds);
+	free(path);
 }
 
-void	do_pipe(t_tokens *tokens, t_shell *shell, int i)
+void	do_pipe(t_tokens *tokens, t_shell *shell, int i, int pid)
 {
-	int	pid = 0;
 	int	fd[2];
 
 	if (pipe(fd) == -1)
 		;
-	pid = fork();
 	if (pid == 0)
 	{
 		if (i == 0 && shell->fd_in != STDIN_FILENO)
@@ -45,6 +50,5 @@ void	do_pipe(t_tokens *tokens, t_shell *shell, int i)
 		close(fd[0]);
 		prepare_exec(tokens, shell);
 	}
-	parent_process(pid, fd);
-	waitpid(pid, NULL, 0);
+	parent_process(fd);
 }
