+++ src/main.c
@@ -40,110 +40,6 @@ static void	init_shell(t_shell *shell, char **envp)
 	shell->original_stdout = dup(STDOUT_FILENO);
 }
 
-char	*get_var(char *token)
-{
-	int		i;
-
-	i = 0;
-	while (token[i])
-	{
-		if (token[i] == '$' || token[i] == '\"' || token[i] == '\'')
-			break ;
-		i++;
-	}
-	token[i] = '\0';
-	return (token);
-}
-
-char	*skip_quote(char *token, const char *quote_type, bool *sq, bool *dq, bool *first_encounter)
-{
-	char	*ret;
-
-	ret = ft_strdup(token);
-	if (quote_type[0] == '\'' && *sq == true && *first_encounter == false)
-		*sq = false;
-	else if (quote_type[0] == '\"' && *dq == true && *first_encounter == false)
-		*dq = false;
-	*first_encounter = false;
-	ret++;
-	return (ret);
-}
-//TODO: free string_joins
-t_tokens	*handle_quotes(t_tokens *tokens, t_shell *shell)
-{
-	t_tokens	*ret;
-	char	*token;
-	char	*temp;
-	char	*envp_var;
-	bool	dq;
-	bool	sq;
-	bool	first_encounter;
-	int		i;
-
-	ret = tokens;
-
-	sq = false;
-	dq = false;
-	first_encounter = true;
-	while (tokens && tokens->token)
-	{
-		sq = false;
-		dq = false;
-		first_encounter = true;
-		token = tokens->token;
-		i = 0;
-		while (token[i])
-		{
-			first_encounter = true;
-			if (token[i] == '\'' && dq == false)
-			{
-				sq = !sq;
-				char *trimed = skip_quote(&token[i], "\'", &sq, &dq, &first_encounter);
-				token[i] = '\0';
-				token = ft_strjoin(token, trimed);
-				continue ;
-			}
-			else if (token[i] == '\"' && sq == false)
-			{
-				dq = !dq;
-				char *trimed = skip_quote(&token[i], "\"", &sq, &dq, &first_encounter);
-				token[i] = '\0';
-				token = ft_strjoin(token, trimed);
-				continue ;
-			}
-			if (token[i] == '$' && sq == false)
-			{
-				if (token[i + 1])
-				{
-					envp_var = get_var(ft_strdup(&token[i + 1]));
-					temp = ft_strdup(&token[i + ft_strlen(envp_var) + 1]);
-					token[i] = '\0';
-					// Free token after strjoin
-					token = ft_strjoin(token, handle_expander(shell->envp, envp_var, shell));
-					i += ft_strlen(handle_expander(shell->envp, envp_var, shell));
-					token = ft_strjoin(token, temp);
-					free(temp);
-					free(envp_var);
-				}
-				else
-					break ;
-				if (!*token)
-				{
-					tokens->token = token;
-					tokens->type = SKIP;
-					return (ret);
-				}
-				continue ;
-			}
-			else
-				i++;
-		}
-		tokens->token = token;
-		tokens = tokens->next;
-	}
-	return (ret);
-}
-
 int	main(int argc, char **argv, char **envp)
 {
 	t_tokens	*tokens = NULL;
